Adds MonoGame.Extended to read Tiled map
Adds:
- MonoGame.Extended
- MonoGame.Extended.Content.Pipeline
- MonoGame.Extended.Tiled

Refrence to "MonoGame.Extended.Content.Pipeline.dll" made in Content.mgcb

Note that in "Content.mgcb" a refrence to "MonoGame.Extended.Content.Pipeline.dll" must be (and has been) made.
For UWP apps this .dll file resides in the user folder, and not the project folder.
Therefore the username "BjAlv" must be replaced with your own username to make it run on your own computer.
Monogame does not seem to support %USERPROFILE% in the reference path at this moment.