Adds simple map, and code to render/draw it
The map is created with Tiled (TiledEditor).
When creating the map it is important to choose "build into map" for the tileset, since this is the only option supported for version 1.1.0 of Monogame.Extended.
The map (.tmx), and the image file (.png or other image format) of the tileset must both be added to "Content.mgbc". The "build" button can then be pressed.

This commit only contains an initial test map, since the main purpose was to find out which setings were required for it to work.
At the current time the documentation for Monogame.Extended is severely lacking.

Also note that in "Content.mgcb", the interpreter and processor is listed as "invalid".
This is due to a bug in MonoGame 3.7, and will be fixed in Monogame 3.8. However 3.8 is not out as a stable release yet. But building manually seems to work just fine regardless.