Adds picking-code for sprites + virtual move logic
Picking logic (Picking2D.cs) is called from / implemented in the public Update method in SpriteClass.cs.
The movement logic (used to control sprite when it is not picked up) is put in a protected virtual method (in SpriteClass.cs) so that it can be overridden by any inheriting class.