Changes HUD positions to be relative to windowsize
Subcribes to event so that positions get updated when window is resized.

The code for changing position on change of DPI is also in this commit, but it is commented out.
It should be uncommented and committed again if it turns out that the code is nessecarry.
If not, then it should be deleted and commited.