Fixes car turning round when wheelspeeds were zero
Checks if wheelspeeds are equal, and if so, does not calculate angularVelocity.
It will stay at 0, which is what it gets initialized at.

Note:
This commit just fixes a glaring bug (car turning arround on the spot, when it should stand still).
The Drive2() methot is still not fully tested, nor fully developed.