Removes Monogame.Extended (Tiled map importer)
Monogame extended was used for importing maps.
However the API documentation was poor / severely lacking. And it was ovely complicated requiring a dll refrence in the content pipeline tool, which for UWP apps is kept in the user folder. This was causing issues for people pulling the code to other computers.

Monogame.Extended are currently working on a 2.0 release which is supposed to improve the API significanly, but we do not have time to wait for that, and since we only use it for importing maps we have decided to go for a more specialized library for parsing Tiled maps instead.

Monogame.Extended is removed and will be replaced by TiledSharp in a later commit.