Adds code for drawing a map using TiledSharp
TiledSharp does not have build in code for drawing the map, only for parsing the TMX file to an object.

This code makes the assumption that the maps will be created with the floor layer at the bottom, and then the wall layer above that.

Note:
There is currently an issue with scaling that for some values the map will get visible grid lines (see comments in LoadContent() method inside SimulatorGame.cs for more details).