Adds class for collecting Lidar Distance Data
This class will start a separate thread where it:
- Utilises LidarPacketReveiver to receive packets from the Lidar.
- Then uses LidarPacketInterpreter to interpret the data packets.
- It then fires an event, sending out the collected data to all its subscribers.

The class is internal as it is meant to be an implementation detail for the assembly where all the LidarClasses resides.
The end user will be using LidarDistance (which will be a subscriber) to get the distances.
The end user can also use the LidarPacketReceiver to collect untreated data packets from the LIDAR.