Adds Interface IDistance for distance measurements
The interface ensures that the all who implements it at least contain a Fwd, Left and Right property as well as Min and Max range, and Resolution.
The intention is that the Distance class will accept classes implementing IDistance in its constructor, and will blend all sensor data it receives.