Adds test code for Lidar in Console Test App
This project is intended for testing parts of the VehicleEquipment library.
This commit contains some early version of tests for LidarDistance class and LidarDistanceCollector.
(And also adds a refrence to the VehicleEquipment library).

Testing the LidarDistance class would also test all the classes below, such as LidarDistanceCollector, LidarPacketInterpreter, LidarPacketReceiver etc.

The purpose of also testing the lower classes seperately (e.g. LidarDistanceCollecor) is to narrow down on the root cause, if the upper class exibits any issues.