Changes class from static and implements interface
LidarPacketReceiver was previously a static class.
This commit changes it to normal class which implements the interface ILidarPacketReceiver, to enable dependency injection (which in the future may give the option of choosing between using a SimulatedLidarPacketReceiver instead of the LidarPacketReceiver).

Note: This is commit will probably break the code (other commits will follow that are also required).