Refactoring / change in how Lidar classes work
Summary:
Changes from packet collecting in separate thread and event to create loose coupling.
Changes to async packet collecting (single thread) and uses DI (interface passed as argument to LidarDistance constructor) to create loose coupling.

Changes TO - details:
LidarDistance now utelizes an async method in a sort-of 'fire-and-forget' way, where it utilises its packetReceiver (which LidarDistance gets as an ILidarPacketReceiver argument in the constructor) to collect packets, and then uses the static LidarPacketInterpreter to interpret the packets.
The Distance will periodically update via the use of the async method.
NOTE:
further investigation must be made to see if this way of running the async method at intervals is an appropiate way to solve the issue.

The event and LidarDistanceCollector class is no longer needed, and therefore deleted.

Changes FROM (previous) - details:
LidarDistanceCollecor.cs running a thread where it was using static LidarPacketReceiver.cs to collect distances, then used LidarPacketInterpreter.cs to iterpret the raw data.
It then fired an event, which was subscribed to in LidarDistance.cs, which then got updated distances.