Changes distance over range calculation logic
Checks if range (fromAngle, toAngle) spanns zero, and then enters loop (instead of checking on each iteration).

The new logic does not iterate whole Distance collection, but instead list is first sorted (in LidarPacketInterpreter), and then in LidarDistance the fromAngle is found by searcing, and then we iterate until toAngle is found.