Adds comment with bug-reminder.
Adds a BUG reminder comment on the Median calculation type inside the PerformCalculation() method.
The Sort() functions sorts by angle (which is what that method is supposed to do).
However for the Median calculation we actually wish to sort by distance instead.

The comment was entered as a reminder to modify the code to a custom sort once the code arround required for testing it is created.