Adds IoC-containers in App.xaml.cs for when on IoT
Registers the containers that is to be used when code is detected as running on an IoT device (i.e. the physical car).

Note:
This code has not been tested yet, as that requires the actual car (and more code), neither of which was available at the time of this committ.