Includes Prism.Core ase a PackageRefrence
This is included to be able to inherit from BindableBase on the models in the VehicleEquipment library  (for simpler INotifyProportyChanged syntax).
And also to access INotifyDataErrorInfo and possibly other useful code.

Later in the project it should be considered replacing this refrence with something else that gives similar functionality (possibly just writing the spesiffic methods we need, if we end up using relatively few methods from the Prism.Core library).

To see what Prism.Core contains, see:
https://github.com/PrismLibrary/Prism/tree/master/Source/Prism