Changes ultrasonic interface and model
Ultrasonic inherits from BindableBase (for simplified propery changed notifications).

Removes the DistanceLeft() etc. methods and replaces them with properties.
Changes logic so that each request does not try to read from microcontroller, but will instead only get new data from the microcontroller if the last reading is older than the selected "PermissableDistanceAge".
The check and update of the distances is done in the UpdateDistanceProperties() method.

The UpdateDistanceProperties() method contains a lock, in case two threads thould attempt to update at the same time. Once the first thread has updated the fields, the second thread will immediately retun the value stored in the fields, instead of trying to collect new data from the microcontroller again.

The IUltrasonic interface was changed to correspond with the changes in the Ultrasonic class.

The GetNewDistances() is mostly the same as it originally was. It has just been renamed from the old method name Distance()