Changes encoder to assemble received bytes to ints
The microcontroller code has also been changed to correspond with this, but only in the test repo.
Changes will be made in microcontroller code in this repo as well at a later stage.

Also sets GetEncoderData() to private, and adds a public Collect... method that is intended to be used for polling the encoder for new distance data.

Also performs som minor renames, and the TimeAccumulated is now milliseconds instead of seconds.