Changes package version from 6.2.3 to 6.1.9
Reduces package version of Microsoft.NETCore.UniversalWindowsPlatform from 6.2.3 to 6.1.9 in Communication library (6.1.9 is the highest available version in the Application project (main UWP project)).

Prior to this change we were getting the following error once the program was starting on the IoT device:
***
system.typeloadexception: 'could not load type 'system.runtime.compilerservices.iasyncstatemachine' from assembly 'System.Runtime, Version=4.2.1.0, Culture=neutral, ...
***

This occured when trying to await async methods in the VehicleCommunication classes SetUpCommunication() method.

Reducing the package version to correspond with the main application using this library seems to have fixed the issue.