Adds Helper project with ArrayConverter
The ArrayConverter is used in the VehicleCommunication class in the Communication project.
But since the Communication project is a Universal Windows class library (required for access to I2c), the classes inside it can't be tested in a normal unit test project.

For that reason, and also to increase reusability, if another usecase should arrise, the ArrayConverter class has been placed in a .NET Standard library (Helpers project).

A corresponding unit test project has also been created, and unittests written for the ArrayConverter class.