Adds PowerViewModes for power and system control
Wraps the IPower object, so that the View can access it.
For convenience there is also added a method which sets all Power properties to false (powers down all equipment).

Also supplies methods for ExitApplication, RestartSystem and ShutdownSystem, which should ONLY be used if the system is runnin on the IoT device.
Therefore a Visibility property 'VisibleIfRunningOnIoT is also available, so that the View can hide the buttons at runtime, if code is running on a non-IoT system.