Sets constant size for byteArray
The byte array that is sent to the microcontroller should be of a pre-determined size, and has to be agreed upon by both sides.
This commit fixes a bug where the byteArraySize was set dynamically based on what was beeing stored in it. It has now been rectified to instead be 23 bytes (set via constant DataTransmitSize in the VehicleCommunication class).