Fixes issue with slider reset when navigating away
Previously a new collection was returned (asynchronously) every time the page was entered.
This caused the sliter to reset, although a thread was still running.
Activating the slider again would cause a new thread to start running the same logich which is allready running (two logics trying to control at the same time).

Now the ExampleLogics collection in initialized only once, when the page is first accessed.
(Note that with these changes, this is no longer done asynchronously).

This means the slider stays put, and can therefore be stopped even after having navigated away from and back to the page.
It is no longer possible to run several threads of the same control logic at the same time.
(However it is still posible to start a different control logic at the same time).