Fixes bug causing multiple ExampleLogic threads
When exiting and entering the ExampleLogic view, while a control logic was set to running, the RunExampleLogic was called again with the same value as before (true).
This caused the if-test to pass, and a new thread was started.

Bug was fixed by returning immediately if the value has not changed since last set-request.
Only one thread is now running, even when exiting and re-entering the page.