Adds threadsafe property changed class
Creates a helper for raising property changed notifications from non-UI threads.
Classes wishing to use this function should inherit from this class.

Note that the default constructor assumes that the object is created from the UI thread.
If that is not the case, a Synchronication context may be passed in through the constructor.

The starting point for this class was some of the methods from Prisms BindableBase code.
This has been rewritten to allow for synchronization against UI thread, and also to use a boolean to choose if PropertyChanged should be raised or not.

(This commit also contains an Annotations.cs file which was autogenereted by Resharper. This is used for intelisence etc. in conjunction with INotifyPropertyChanged)