Moves ArrayConverter to VehicleEquipment project
It was moved from Helper-project to VehicleEquipment-project since the ArrayConverter class depends on the VehicleEquipment projects Devices and MessageCode enums, and also the VehicleDataPack class.
The Helpers project should contain only classes that may be used everywhere, and does not depend on any other projects.

(Prior to this move, VehicleEquipment would not be able to use classes in the Helpers project, since it could not refrence it, when the Helpers project was already refrencing the VehicleEquipment project).