Makes Ultrasound-model raise selective INPC
Makes IUltrasound interface extend INotifyPropertyChanged, and Ultrasounc class inherit from Helper-class ThreadSafeNotifyPropertyChanged (which in turn implements INotifyPropertyChanged).

This allows the Ultrasound model to raise property changed notifications.

This way the View can bind directly to the properties (by exposing the model in the ViewModel). However it will only raise the property changed event for the distance properties IF we set the RaiseNotificationForSelective to true, so that the ControlLogic doesn't raise many notifications when we aren't looking at a view where we actually observe those values.