Changes ViewModel to leverage the INCP in Model
The RaisePropertyChanged no longer gets raised at intervals in the ViewModel, but rather the Model will raise PropertyChanged when there has actually occured a change (and selector bool is true).

The selector bool gets set to true automatically when entering the view (if Power is on), and will be set to false again upon leaving view.