Adds exception handling to Ultrasonic sensor
Any exception that occures while trying to collect data gets stored as a string in a Message Property.

This gets displayed in a textblock on the Ultrasonic View.
Tapping the textbox clears the message and error.

It will not try to collect new values until error has been acknowledged (returns NaN for all distances).