Adds INotifyPropertyChanged to LidarDistance
ILidarDistance interface extends INotifyPropertyChanged.
In LidarDistance class it is implemented by inheriting from ThreadSafeNotifyPropertyChanged (which is a class in the Helpers project).

The interface also has a RaiseNotificationForSelective property, which exposes the corresponding property from ThreadSafeNotifyPropertyChanged class.
(Gets used for selecting weather the selective property should raise notifications or not).