Adds and uses custom ObservableCollection
This collection ensures that there will not be added any duplicates.
It also will raise CollectionChanged notification on the UI thread, even if item is added or removed from a different thread.

NOTE:
This class is not inherrently thread safe, and must therefore be locked appropiately where used with multiple threads (will be done at a later stage).