Adds BUG comment reminder in GetDistancecInRange()
Noticed issue when we had no valid values in left sector of lidar.
If perfroming a calculation that spanns zero (i.e. Fwd), it will return List with only a float.NaN, even if there are valid values on the right hand side.
To be fixed later when improving the calculation logic.