Utilizes MinRange from LidarDistance class
Makes the LidarInterpreter relie on the MinimumRange selected in the LidarDistance class (given as an argument), istead of having its own internal one.

Also adds MinRange and MaxRange to the interface, so that they can be bound to.
(Note: The MaxRange is not actually used for anything as of yet).