Adds optional boolean to force new wheelspeed
if it is true (default) the new setpoint will not be sent if it is the same as the previous setpoint.
If set to false, it will be sent regardless.

Intended usage is to force a stop command to be sendt, even if the system believes that it has allready been sent.