Adds error message properties and methods
The HasUnacknowledgedError property will be used used for putting a red frame arround the shell button when an error is active, and from preventing re-start of control logic until error has been acknowledged.

ClearMessgage() method will be used for removing the message stored in ErrorMessage property, and to allow the control logic to be started again.