Adds error message box for Power class
The message box on the PowerPage is used for displaying error messages from exceptions to the user.
It can be acknowledged/cleared by tapping the message.

The Power class utilises:
- String property "Message" for storing the error message.
- ClearMeassage() method for clearing the message and setting HasUnacknowlededError proerty to false.
- Boolean HasUnacknowlededError property for keeping track of whether an error is active or not (blocks switching power state if true).

NOTE:
Exception handling for state-switcing has just been added for the Lidar as of this commit (TODO comment added).
Exception handling for opening of pin has been added for all (via OpenNewPowerPin() method).