Adds shell-button colors/visibility based on state
If an example logic is running, or a sensor is powered on, the respective shell-button will turn green.
If it has an error, it will get e red frame.

In addition, if an example logic is running, the shell button for student logic will be hidden.
(Student Logic is not yet implemented, but once it is it will also hide the Example Logic if a Student Logic is running).