Instanciates 'Simulated coms.' when not run on IoT
These will get injected into the constructors of LidarDistance, Ultrasonic, Encoder etc. at runtime, if the code is not running on an IoT device.

Also adds a private filed of type SimulatorAppServiceClient and passes it as an parameter to the simulated classes. (The SimulatorAppServiceClient is intended to be used for the actual communication against the Simulator).