Sets up Encoder (mock)simulation on simulator side
Currently it doesn't return the distance traveled as it is supposed to, but rather returns the speed at that spesiffic moment in time. A TODO comment has been added in the project as a reminder to be rectified when we have time.

(EncoderRight has not been tested, since this is not implemented in the main program, nor on physical hardware yet).