Uses the IsRunningOnPysicalCar bool instead of checking again itself
This was changed so that if we later on decides to run a different type of check to see if it is running on the physical car, then we only have one place in the code where we have to change it (in the App.xaml.cs file).
Currently we only are cheking if it is an IoT device.