Adds additional ultrasonic sensor (at foreward)
Previously there were only 1 sensor in the front. Now there are two. One located to the left of the centerline, and one to the right of the centerline.

The Fwd property allways returns the smallest of the two.
The FwdLeft and FwdRight properties can also be used to read the spesiffic sensor, instead of the smallest of the two.