Puts distance data in horizontal StackPanel
This was done to prevent the data overlapping, and to make positioning it easier.

Previously both distance list and directions were inside the same relative panel, where the directions were centered in the relative panel, causing inefficient use of space and overlapping.

This was changed to instead have two RelativePanels inside a horizontal StackPanel.