Imroves Wheels view layout (not complete yet)
The encoder part is currently organized in nested StackPanels in such a way which makes it difficult to make changes to the layout.
TODO comment has been added to try and replace with RelativePanel.

We will wait with those changes until the Encoder class has been modified (to allow for a second encoder on the right hand side).