Removes speed reduction from ExampleLogic L1b
This was removed due to there is now a safety function coded into the Arduino which will automaticalle reduce speed if it gets too close to an obstruction.
Doing this also in the control logic then caused the car to end up driving very slowly.

Also increases distance from obstruction in front of car before it rotates away from obstruction from 0.2 meter to 0.5 meter.