Moves Left Encoder into a common Encoders class
Previously the code was set up for use of only one encoder (for left wheel).
The new Encoders class has two instances of Encoder (one for Left and one for Right).

The new design is a result of the decition to have one microcontroller for each encoder (and not run both encoders on one microcontroller).