Changes Encoders CollectionInterval to int
This commit fixes the issue where if the user entered a number into the seconds field in the textbox, it would be ignored (previously it only looked at the millisecond part, which at max could be 999).

Also limits the smallest value the user is allowed to set (againts a private constant).