Additional fix to multiple collectors running bug
Sometimes when togle switch was pushed, the setter was accessed two times in quick succession with the same value (true) causing the collector to start twice.
This same issue could also be caused if property were to be set from ControlLogic to the same value it allready was.

This issue has now been fixed.
(setting collectorTokenSource to null, and checking in the if-block is probably not neccesary, but was added as extra safety measure).