Changes properties to private and AvgVel from calc
Sets propoerty setters to private since there is no reason for them to change from outside the class.
Therefore a method is also created to reset TotalDistance inside the Encoder class, so that Encoders class doesn't have to call the properties for Left and Right encoder directly.

Average velocity property has been changed from caluclated to one that gets set when new data is received.
(Also adds a lock arround the code for calculation)
The reasons for this are:

- The only time it is nessecary to calculate the distance is after new data is received.

- If it were calculated on request (which could happen from a different thread) there is a slim chance that the Distance has been updated, but the Time has not, and that we therefore would calculate an incorrect value.