Replaces Lidar-settings TextBoxes with ComboBoxes
This was done to make them easier to use on the vehicles touch screen.

TODO/BUG reminder has been added in comments to look into making them editable (1809) in a robust manner.

The LidarDistance MinRange and MaxRange has been changed from Float to Double, since x:Single did not seem to be available in the XAML for our UWP app (despite the .NET Framework having it https://docs.microsoft.com/en-us/dotnet/framework/xaml-services/built-in-types-for-common-xaml-language-primitives ).