Separates control logics for turning and steering
This is now separated into two control logics.
One which only turns the vehicle towards the greatest distance.
And one with steers towards the greatest distance (ignoring obstacles).

Both are intended to be visible for the students, as they just contain simple code demonstrating how the LIDAR could be used for steering.