Changes direction-to-turn logic and increases limit fwd
Note that using 0.3f was working fine when testing (without safety function active).
But to avoid interference by safety function once it gets fixed, the distance was increased to 0.5.
This had the side effect of the ultrasound now detecting the door-thresholds, and therefore no longer able to pass over them.
We are considering angeling the sensors upwards slightly to avoid this.