Adds ScaleToHighDPI method without rounding
The rounding was a nessecary addition when scaling the map, to prevent gaps in the rendering of the map on ceirtain scale and screen resolution combos.
However this rounding makes scaling of the car inaccurate. Therefore a new method without rounding was created to be used with scaling of car.