Improves simulated ultrasound by adding offsets
Also creates a method for calculating Ultrasound distance based on the measurments in the Simulator.
(e.g. the real ultrasound sensors on the front will not detect the walls on the side if it is driving perpendicular to it, whereas the simulated ones previously would).

Also adds a TODO reminder to check the behaviours of the real ultrasound sensors, and modify simulator against those measurements.