Makes SimulatedLidar Update time scale with cycles
The real packet collector takes longer time to return new values if more cycles are collected.
Uses number of cycles (and constant time for one cycle) to calculate a delay, which will change depending on number of cycles that are collected.

Note: The TimePerCycle constant of 200ms is approximate. Added TODO note to measure update time for real PacketReceiver and then modify constant.