Changes MasterHeader and reduces details margins
The left, top and right margins for DetailsPage (used in Student and Example logic) were reduced, so that more text can fit in the window.

Also the default MasterHeader text in both windows were changed to a more descriptive text ('MasterHeader' is the title displayed on the top of the list).