Fixes slider reset when starting logic with error
Fixes so that the slider no longer stays in the on position when attempting to restart a control logic which has an unacknowledged error. It will now jump back to the off position.
Previously the slider would move to on position, and RunExampleLogic would be set to true (however it did not actually start the control logic again). This would give a false impression that the control logic was running.