Adds buttons for starting extra functionality
Adds button for launching program with extra functionality (like website with webcam stream).

Also adds two sliders for starting/stopping (Async) Socket Server.
(Actual code they should launch is not implemented yet. TODO comments added).

Also takes in sensor objects in the constructor and stores them in private fields, so that they can be passed on to the Socket Server when it is started.