Creates Error class to hold error message etc.
This was done since most of the models and the two control logic bases all requires these fields/methods.
Extracting this out to a separate class which gets instantiated in each of those models instead will keep the code DRY and the other classes cleaner.

It also makes it easier if we wish to add logging at a later stage, since this then can be easily placed in the Error class.