Improves error-message display/code for Ultrasound
Replaces error-related fields with Error class in the Ultrasonic Model.

Also improves the displaying of error message in the view/page.
This new display is the same as the one developed for the ExampleLogics, with red text and buttons for toggling details or clearing error.
Also added a TODO reminder to try and find a good solution for scrolling text to prevent data being pushed off screen if the exception message is too long.