Fixes issue where error message extends off screen
Adds ScrollViewer to Ultrasonic pages error message display to fix issue where the buttons (and part of the error message) were beeing pushed of screen if the error message was very long.
The ScrollViewer MUST have a MaxHeight (or Height) set, or else the ScrollViewer will simply resize to be larger than the content, and therefore not work.

It is now possible to zoom the text in/out within set limits, and to scroll the text vertically.