Improves error-message display/code for Wheels
Replaces error-related fields with Error class in the Wheel Model and Encoder(s) Model.

Also improves the displaying of error message in the Wheel view/page.
This new display is based on the one developed for the Encoder, with red text and buttons for toggling details or clearing error.
The error message is scrollable if the message is longer than what will fit in the ScrollViews MaxHeight.
It can also be zoomed within set limits.

The red error frame arround the shell button has been fixed so that it will light up on alarm from Wheel OR Encoder.