Improves error-message display/code for LIDAR
Replaces error-related fields with Error class in the Lidar Model.

Also improves the displaying of error message in the Lidar view/page.
This new display is based on the one developed for the Encoder, with red error text, and buttons for toggling details or clearing error.
The error message is scrollable if the message is longer than what will fit in the ScrollViews MaxHeight.
It can also be zoomed within set limits.