Adds code to stop ControlLogics on sensor error
This has NOT been added to the ControlLogicBase classes, but is instead placed inside the individual control logics.

This was done so that the people making the control logics have the freedom to decide how they will handle the sensor errors. (Maybe they wish their control logic to switch to using a different sensor type, if their prefered sensor fails, instead of stopping the control logic).