Adds prefered Launch View Size
The prefered size is set to 800x480 since that is the resolution of the RaspberryPI screen.
(On the RaspberryPi it will run full screen regardless of this setting, but it was set to this resolution so that it will look the same when launched on a PC).

The vertical may have to be increased since it probably includes the vertical title-bar, which is not present when running on the car/RaspberryPi.