Moves Power from separate class to Sensor Models
Removes Power class (where all GPIO pins / Power on-off were), and instead inject a IGpioOutputPin into the individual sensor models.
This will give us more flexibility in the control of the sensors, since the sensor model can be aware of its power state.

This commit makes only the changes required so that we can take advantage of the increased flexibility in future commits. It does not yet offer any different experience to the end user.

Note that these changes hasn't been tested on the physical vehicle at the time of this commit (only tested against simulator).