Makes AutoCollect stop when 'Display Data' is off
Prior to this commit, the Display Data slider could be turned off, but the AutoCollecting of Ultrasound distances (RefreshUltrasonicContinously) could still be running in the background.

Makes the RaiseNotificationForSelective virtual so that it could be overridden in the Ultrasonic class, and added code to switch off AutoCollect (RefreshUltrasonicContinously) when the RaiseNotificationForSelective is switched off.

Also switches off the AutoCollect it there is an unacknowledged error.