Changes GpioOutputPin to also handle Input
This involves changing names on:
IGpioOutputPin --> IGpioPin
GpioOutputPin --> PhysicalGpioPin
SimulatedGpioOutputPin --> SimulatedGpioPin
MockGpioOutputPin --> MockGpioPin

Also makes changes inisde PhysicalGpioPin and SimulatedGpioPin to enable object to be constructed with pin in either an Input or an Output state.

Currently the ValueChanged event is not available (since the GpioPin object is kept private, and no methods expose it yet).