Changes Ultrasonic class to utilize InterruptPin
Prior to this commit the decision on whether to Collect new distance from microcontroller, or reuse the old data, was based on time since last collection.

If autocollection was activated, it would also collect at set intervals (running in a separate task/thread).

This commit changes so that it will instead base this on a GPIO pin configured as input.
The microcontroller sets the pin high when there is new distance data available, and sets it to low when the data has been collected by the Main Program. Main program will reuse old distance data if this pin is low.

Changes the autocollection to instead trigger on ValueChanged event for the pin.