Adds SocketServer for external communication
This socket server leverages async. Not fully implemented yet.
Current state:
- When StartServer() is called, it goes into a listening state.
- Once connection is received, it reads until it gets a new line character.
- It then returns the same message it received back to the owner.
- Socket shuts down (must be restarted to receive new messages).

Verified that it is working with a simple consol app sending messages to it.
Also testen by running this code on the SensorVehicle and sending message with Putty (connection type raw).

The intended use case for this socket is receiving wheel commands for external unit, responding with sensor data upon request.