Fixes issue with message not beeing sent to server
Adds newline to message beeing sent if payload does not allready contain it (server expects newline to end the message).
Previously no newline was beeing sent, causing server to not read the message until the PrintWriter was closed.

Replaces PrintWriter with a BufferedWriter, and also adds a BufferedReader to read the response message from the server.

Also makes socket close after EXIT_CONFIRMATION is received from the server.
However code for closing socket must be improved further (i.e. closing socket in the event of a failure / no server response).