Adds possibility to stop control logic via socket server
If any control logic is running the only wheel command that can be set via Socket Server is zero speed. Setting zero speed will cause control logic to stop (and confirmation of this sent as a response).
For any other wheel speed command, a response will be sent informing that a control logic is running, and must be stopped first.

If no control logic is running, wheel-commands can be given as normal.