Adds ability Stop/Restart ControlLogic via SocketServer
Server side has been set up to accept "components" StopControlLogic and RestartControlLogic.
If StopControlLogic command is received and no control logic is running, it will simply send a stop command to the wheels instead.

The android app now has two buttons on the move page and on the data page.
- Stop button sends stop request for any active control logic.
- Restart control logic sends request to restart the previously stopped control logic.
(logich for which control logic to restart occures on the server, not on the android app)