Changes UltrasoundInterruptPin from 5 to 19
This is to correspond with the change in physical connection, where it was moved from GPIO5 (which has default Pull High), over to GPIO19 (which has default Pull Low.

It is desirable with Pull Low, since the program will request new distance data from ultrasound microcontroller when the pin goes high.