Powers down wheels when error occures (IF moving)
Changes catch block so that power to micro-controller for wheels gets switched off if an exception occures (during communication when setting wheel speed) - AND the last successfull command was not 0.

If last successfull command was 0 (or no command has been sent yet), the microcontroller remains powered on. This was done to avoide fustration when faultsearching in the case of a communication error via the I2c.