Removes unused constructor parameters
The StudentLogicViewModel and ExampleLogicViewModel does not use the IWheel, ILidarDistance and IUltrasonic it requested in the constructor.
These are instead extracted directly in the constructors for StudentLogicService and ExampleLogicService (which are themselves registered as unity-containers).

Also removes the new InjectionConstructor(...) for registering the ISocketServer/SocketServer as a unity container, since the parameters it is using are all registered containers itself, it will get resolved without having to explicitly inject them in App.xaml.cs.