Adds Enum-to-Visibility converter
This may be used for binding Visibility properties in XAML to Enums.

Note that the converter will generally be declared logaly in the XAML where it is to be used (as opposed to globaly in the App.xaml, since we must specify an enum type for it when we declare it.
But if we want to use it for the same enum type im many XAML files, we can still declare it globally in App.xaml instead.

USAGE SYNTAX:
**  On the top layer of XAML declare:  **
<Page.Resources>
    <converters:EnumToVisibilityConverter x:Key="EnumToVisibilityConverter" EnumType="helpers:MyEnumType" />
</Page.Resources>

** Usage in binding **
Visibility="{x:Bind ViewModel.MyEnumHoldingTheState, Mode=OneWay, Converter={StaticResource EnumToVisibilityConverter}, ConverterParameter=TheStateThatShouldSetVisibility}