Adds download button for simulator on InfoPage
The download buton is only visible if no simulator was found on the PC, and the SensorVehicle Main app is connected up against MockData.
The download button will launch the Microsoft Store page where the simulator may be downloaded.

We use the EnumToVisibilityConverter to set the visibility on the download button.