Adds new catch that doesn't shut down SocketServer
The new catch block for COMException catches exceptions which seems to happen some times when the vehicle changes accesspoint while being controlled by android app.
When this exception is caught a stop command is sent to the wheels, but the SocketServer remains running so that the user can reconnect to the socket server.

(When a general Exception occures that is not handled elsewhere, the SocketServer gets shut down to be on the safe side).