Adds popup box to notify when network connection changes
This will not triger when it just changes access point, but does trigger if wifi gets disconnected.
Issue where IP address gets lost has not reoccured after listening to this event, so we are not sure if that will trigger it.

Also adds ToDo reminder to implement automatic reconnect (if we get popup every time we get network issues).